app/requirements.txt
flask
rembg
pillow
gunicorn
onnxruntime